refactor(bmdashboard): modularize material insights controller
Split controller into calculations, handlers, and composition layers
to resolve ESLint violations and improve maintainability.
